deployment/requirements.txt

pandas
numpy
streamlit
seaborn
matplotlib
scikit-learn==1.3.0
plotly
joblib
image
pillow
tensorflow